+++ typeClassDefs/wb_vision_control_status.c
@@ -758,12 +758,12 @@ struct wb_vision_control_status* wb_vision_control_status_from_string(struct wb_
             }
         }
         if (endKey >= startKey && endKey - startKey < length) {
-            strncpy(key, str + startKey, (endKey - startKey) + 1);
+            strncpy(key, str + startKey, ((size_t)(endKey - startKey) + 1));
             key[(endKey - startKey) + 1] = 0;
         } else {
             key[0] = 0;
         }
-        strncpy(var_str, str + startVar, (index - startVar) + 1);
+        strncpy(var_str, str + startVar, ((size_t)(index - startVar) + 1));
         var_str[(index - startVar) + 1] = 0;
         bracecount = 0;
         index += 2;
